program-05
Problem
You are given a permutation with length n. You want to play a game with you friend, Bob, and the rule will be as follows:
• You will choose a subarray [, ] from the permutation. Then, ask Bob to find the maximum element in the rest of the numbers.
You are given a permutation & contains all numbers 1 ton. And, in q queries, each query has two integers / and r.
For each query, you have to help Bob find the maximum value that does not exist in the subarray [1, r]. Note: A permutation is a sequence of integers from 1 ton of length n containing each number exactly once. For example, (1), (4, 3, 5, 1, 2), (3, 2, 1) are permutations and (1, 1), (4, 3, 1), (2, 3, 4) are not.
Input format
• The first line contains two integers n and a denoting the number of elements and the number of test cases. • The second line contains n space-separated integers.
• The next q lines and each line contains two integers 1, r.
Output format
Print the maximum number that does not exist in the subarray [1, r].
Constraints
2≤ n, g≤ 105
1≤aj ≤ n
1 <l, r < n
1l-r+1<n